Initial changes from HuC to HuCC.
Revert as much as possible of the change from "int" to "intptr_t" that
was done in a hurry a few years ago to allow a 64-bit build of HuC.

This also *begins* the process of cleaning up some stuff in the SmallC
codebase that should have been "bool" or something other than an "int".

Removed the remains of Uli's interrupt-handling additions.

Begin renaming some of the macros as part of figuring out how to fix
the code-generation for conditionals.

+++ src/hucc/code.h
@@ -8,7 +8,7 @@
 
 #include "defs.h"
 
-extern intptr_t segment;
+extern int segment;
 
 void gdata (void);
 void gtext (void);
@@ -18,9 +18,9 @@ void asmdefines (void);
 void defbyte (void);
 void defstorage (void);
 void defword (void);
-void out_ins (intptr_t code, intptr_t type, intptr_t data);
-void out_ins_ex (intptr_t code, intptr_t type, intptr_t data, intptr_t imm_type, intptr_t imm_data);
-void out_ins_sym (intptr_t code, intptr_t type, intptr_t data, SYMBOL *sym);
+void out_ins (int code, int type, intptr_t data);
+void out_ins_ex (int code, int type, intptr_t data, int imm_type, intptr_t imm_data);
+void out_ins_sym (int code, int type, intptr_t data, SYMBOL *sym);
 void gen_ins (INS *tmp);
 void gen_code (INS *tmp);
