+++ src/hucc/pragma.h
@@ -10,10 +10,10 @@ void dopragma (void);
 void defpragma (void);
 void parse_pragma (void);
 void new_fastcall (void);
-intptr_t fastcall_look (const char *fname, intptr_t nargs, struct fastcall **p);
-intptr_t symhash (const char *sym);
-intptr_t symget (char *sname);
-intptr_t strmatch (char *lit);
+int fastcall_look (const char *fname, int nargs, struct fastcall **p);
+int symhash (const char *sym);
+int symget (char *sname);
+int strmatch (char *lit);
 void skip_blanks (void);
 
 extern struct fastcall *fastcall_tbl[256];
