+++ src/hucc/pseudo.h
@@ -7,14 +7,14 @@
 
 void dopsdinc (void);
 void dopsddef (void);
-intptr_t outcomma (void);
-intptr_t outnameunderline (void);
-intptr_t outconst (void);
+int outcomma (void);
+int outnameunderline (void);
+int outconst (void);
 void doset_bgpalstatement (void);
 void doset_sprpalstatement (void);
 void doload_spritesstatement (void);
 void doload_backgroundstatement (void);
-void do_asm_func (intptr_t type);
-char *new_string (intptr_t und, char *a);
+void do_asm_func (int type);
+char *new_string (int und, char *a);
 
 #endif
